renamed inner cmd -> subcmd
