+++ cmd/plasmacli/main.go
@@ -2,12 +2,12 @@ package main
 
 import (
 	"fmt"
-	"github.com/FourthState/plasma-mvp-sidechain/cmd/plasmacli/cmd"
+	"github.com/FourthState/plasma-mvp-sidechain/cmd/plasmacli/subcmd"
 	"os"
 )
 
 func main() {
-	rootCmd := cmd.RootCmd()
+	rootCmd := subcmd.RootCmd()
 	if err := rootCmd.Execute(); err != nil {
 		fmt.Println(err)
 		os.Exit(1)
