+++ cmd/plasmacli/subcmd/keys/delete.go
