+++ cmd/plasmacli/subcmd/query/errors.go
