+++ cmd/plasmacli/subcmd/query/info.go
