+++ cmd/plasmacli/subcmd/query/root.go
